move `inject_transit` from SimulatedRainbow to all Rainbows
