+++ chromatic/rainbows/actions/__init__.py
@@ -3,3 +3,4 @@ from .trim import *
 from .normalization import *
 from .align_wavelengths import *
 from .conversions import *
+from .inject_transit import *
