+++ chromatic/rainbows/actions/descriptions.txt
@@ -3,6 +3,7 @@ cartoon | name | description
 🌈🫓😑 | normalize | Normalize by dividing through by a typical spectrum (and/or light curve).
 🌈🍱💇 | trim | Trim away wavelengths or times that contain no good data.
 🌈🚧🌊 | align_wavelengths | Align spectra with different wavelength grids onto one shared axis.
+🌈🪐🚞 | inject_transit | Inject a transit signal into all wavelengths.
 🌈🧮📝 | + | Add two Rainbows together.
 🌈🧮📝 | - | Subtract one Rainbows from another.
 🌈🧮📝 | * | Multiply two Rainbows together.
